fix(deno): address import issues in Deno
